Added (now failing) test for uploading a recording from a collection device
