+++ test/deviceapi.py
@@ -1,11 +1,8 @@
 import json
-import os
-import json
 from urllib.parse import urljoin
 from datetime import datetime
 
 import requests
-from requests_toolbelt.multipart.encoder import MultipartEncoder
 
 from .apibase import APIBase
 
@@ -23,24 +20,6 @@ class DeviceAPI(APIBase):
         if not props:
             props = {}
         return self._upload('/api/v1/audiorecordings', filename, props)
-
-    def _upload(self, url, filename, props):
-        url = urljoin(self._baseurl, url)
-        json_props = json.dumps(props)
-
-        with open(filename, 'rb') as content:
-            multipart_data = MultipartEncoder(
-                fields={
-                    'data': json_props,
-                    'file': (os.path.basename(filename), content),
-                })
-            headers = {
-                'Content-Type': multipart_data.content_type,
-                'Authorization': self._token
-            }
-            r = requests.post(url, data=multipart_data, headers=headers)
-        self._check_response(r)
-        return r.json()['recordingId']
 
     def record_event_data(self, eventData, times):
         eventData["dateTimes"] = times
