+++ test/testuser.py
@@ -247,6 +247,20 @@ class TestUser:
     def get_audio_schedule(self, device):
         return self._userapi.get_audio_schedule(device.devicename)
 
+    def uploads_recording_for(self, testdevice):
+        print("    and '{}' uploades a recording for {} ".format(
+            self.username, testdevice.devicename))
+        props = testdevice.get_new_recording_props()
+
+        filename = 'files/small.cptv'
+        recording_id = self._userapi.upload_recording_for(testdevice.devicename, filename, props)
+
+        # Expect to see this in data returned by the API server.
+        props['rawMimeType'] = 'application/x-cptv'
+
+        return TestRecording(recording_id, props, slurp(filename))
+
+
 class RecordingQueryPromise:
     def __init__(self, testUser, queryParams):
         self._testUser = testUser
